Corrected the C code - pass and print unsigned long long for the step num and get the 0^Inf "stuff" correct.
Add a few longer guesses...

+++ Code/Turing_Machine.h
@@ -53,6 +53,6 @@ typedef struct
 } TM;
 
 extern inline int  step_TM (TM* m);
-extern inline int  step_recur_TM (TM* m, int step_num);
+extern inline int  step_recur_TM (TM* m, unsigned long long step_num);
 extern inline void print_TM(TM* m);
 extern inline void free_TM (TM* m);
